Add RejectAllResponses method in RequestManager
